language fixed

+++ mcd_task/command_actions.py
@@ -4,8 +4,8 @@ import time
 from typing import Callable, Any, Optional, Union, List
 from mcdreforged.api.all import *
 
-from mcd_task.globals import *
-from mcd_task.task_manager import TitleList, TaskNotFoundError, Task
+from mcd_task.constants import PREFIX, DEBUG_MODE
+from mcd_task.task_manager import TitleList, TaskNotFoundError, root, TaskBase, Task
 
 
 # ===============================
@@ -15,9 +15,10 @@ from mcd_task.task_manager import TitleList, TaskNotFoundError, Task
 
 def register_cmd_tree(server: PluginServerInterface):
     def pliteral(literal: Union[set, str]) -> Literal:
+        ltr = literal
         if isinstance(literal, set):
-            literal = list(literal)[1]
-        lvl = root.config['permission.{}'.format(literal)]
+            ltr = list(literal)[1]
+        lvl = root.config['permission.{}'.format(ltr)]
         lvl = lvl if isinstance(lvl, int) else 0
         return Literal(literal).requires(
             lambda src: src.has_permission(lvl),
@@ -25,16 +26,20 @@ def register_cmd_tree(server: PluginServerInterface):
         )
 
     def exe(func: Callable[[CommandSource, str], Any]):
-        return lambda src, **ctx: func(src, **ctx)
+        return lambda src, ctx: func(src, **ctx)
 
     server.register_command(
-        Literal(PREFIX).on_child_error(CommandError, cmd_error, handled=True).runs(info_task).then(
-            pliteral("list").runs(info_task)
+        Literal(PREFIX).runs(list_task).then(
+            pliteral("list").runs(list_task)
         ).then(
             pliteral("help").runs(show_help)
         ).then(
             pliteral("detail").then(
-                QuotableText("titles").runs(exe(info_task)))
+                QuotableText("title").runs(exe(info_task)))
+        ).then(
+            pliteral("detail-all").runs(all_tasks_detail)
+        ).then(
+            pliteral("list-done").runs(list_done)
         ).then(
             pliteral("add").then(
                 QuotableText("titles").runs(exe(add_task)).then(
@@ -59,7 +64,7 @@ def register_cmd_tree(server: PluginServerInterface):
         ).then(
             pliteral("deadline").then(
                 QuotableText("titles").then(
-                    QuotableText("ddl")))
+                    Float("ddl").runs(exe(set_task_deadline))))
         ).then(
             pliteral("player").then(
                 QuotableText("name").runs(exe(info_player)))
@@ -119,22 +124,19 @@ def _player_info_simple(name: str, indent=8, removed=False) -> RTextList:
     """
     return _indent_text(indent) + rclick(
         name, tr("mcd_task.info_player_hover"),
-        "{} player {}",
+        "{} player {}".format(PREFIX, name),
         color=RColor.gold if not removed else RColor.dark_red,
         style=RStyle.strikethrough if removed else None
     )
 
 
-def _task_info_simple(raw_titles: Union[TitleList, str], full=False, sub=False, done=False, idt=None, desc=False):
-    """
-    Task text with click events in a line
-    :param raw_titles:
-    :param sub:
-    :param done:
-    :return:
-    """
-    titles_commands = str(raw_titles)
+def _task_info_simple(raw_titles: Union[TitleList, str],
+                      full=False, sub=False, done=False, idt=None,
+                      desc=False, desc_text: Optional[str] = None,
+                      ddl=None):
+    titles_commands = str(raw_titles).strip('.')
     titles = str(raw_titles) if full else TitleList(raw_titles).tail
+    root.debug("Building title text for {}".format(titles))
     # Indent text
     if idt is None:
         idt = 8 if sub else 4
@@ -156,7 +158,7 @@ def _task_info_simple(raw_titles: Union[TitleList, str], full=False, sub=False, 
     # Task title
     title_text = rclick(
         titles, tr("mcd_task.info_task_hover"),
-        "{} info {}".format(PREFIX, titles_commands),
+        "{} detail {}".format(PREFIX, titles_commands),
         color=RColor.dark_gray if done else RColor.yellow
     )
 
@@ -168,37 +170,62 @@ def _task_info_simple(raw_titles: Union[TitleList, str], full=False, sub=False, 
     ) if not sub else ''
 
     # Description
-    desc_text = root.task_manager[titles].desc
-    description = ''
-    if desc_text != '':
-        description = RTextList(
-            '\n',
-            RText(desc_text, color=RColor.light_gray),
-            rclick(
-                "[✎]", tr("mcd_task.edit_task_hover"),
-                "{} change {} ".format(PREFIX, titles_commands),
-                action=RAction.suggest_command
-            )) if desc else RTextList(
-                ' ',
+    if desc_text is not None or ddl is not None:
+        if desc:
+            description = RTextList()
+            if desc_text is not None:
+                description.append(
+                    '\n', _indent_text(idt),
+                    RText(desc_text, color=RColor.gray), ' ',
+                    rclick(
+                        "[✎]", tr("mcd_task.edit_task_hover"),
+                        "{} change {} ".format(PREFIX, titles_commands),
+                        action=RAction.suggest_command
+                    ), ' ')
+            if ddl is not None:
+                if desc_text is None:
+                    description.append('\n', _indent_text(idt))
+                description.append(
+                    rclick(
+                        formatted_time(ddl), tr("mcd_task.set_ddl_hover"),
+                        "{} deadline {}".format(PREFIX, titles_commands),
+                        action=RAction.suggest_command, color=RColor.gold
+                    )
+                )
+        else:
+            description = RTextList(
                 rclick(
                     "[...]", tr("mcd_task.show_desc_hover"),
-                    "{} detail {} ".format(PREFIX, titles_commands),
-                    action=RAction.suggest_command
+                    "{} detail {} ".format(PREFIX, titles_commands)
                 )
             )
+    else:
+        description = ''
 
     return RTextList(
-        indent, button, title_text, rename_task_button, description
+        indent, button, ' ', title_text, ' ', rename_task_button, description
     )
 
 
 def _task_info_flexible(title: Union[TitleList, str], sub=False, idt=4):
     title = TitleList(title)
-    target_task = root.task_manager[title]
-    text = RTextList('\n', _task_info_simple(title, sub=sub, done=target_task.done, idt=idt, desc=not sub))
+    root.debug(f"Trying to detail {str(title)}")
+    target_task = root.task_manager[title]  # type: TaskBase
+    desc = target_task.description
+    text = RTextList(
+        '\n',
+        _task_info_simple(
+            target_task.full_path(), sub=sub, done=target_task.done,
+            idt=idt, desc=not sub, desc_text=desc if desc != '' else None,
+            ddl=target_task.deadline if target_task.deadline != 0 else None
+        )
+    )
+    if len(target_task.sub_tasks) == 0:
+        return text
+    root.debug("sub_tasks.rue")
     for t in target_task.sub_tasks:
         text.append(
-            '\n', _task_info_flexible(t.full_path(), sub=True, idt=idt + 4)
+            _task_info_flexible(t.full_path(), sub=True, idt=idt + 4)
         )
     return text
 
@@ -227,34 +254,36 @@ def formatted_time(timestamp: float, locale: Optional[str] = None) -> str:
 
 def _add_task_button(title: str = None):
     # Add button
-    title_path = '' if title is None else '{}.'.format(title)
+    title_path = '' if title in ['', None] else '{}.'.format(title)
     return rclick(
         '[+]', tr("mcd_task.add_task_hover"), '{} add {}'.format(PREFIX, title_path),
         action=RAction.suggest_command, color=RColor.red, style=RStyle.bold
     )
 
 
-def _info_task(source: CommandSource, title: Optional[str] = None, done=False) -> Optional[RTextList]:
+def _info_task(title: Optional[str] = None, done=False) -> Optional[RTextList]:
     target_task = None
     if title is not None:
         title = title.strip(".")
         try:
             target_task = root.task_manager[title]
         except TaskNotFoundError:
-            task_not_found(source)
-            return None
+            target_task = None  # type: Optional[TaskBase]
 
+    root.debug("_info_task: target_task = {}".format('None' if target_task is None else target_task.full_path()))
     text = RTextList()
 
     # Task tree
     if target_task is None:
-        for t in root.task_manager.split_sub_tasks_by_done()[0 if not done else 1]:
-            text.append('\n', _task_info_simple(t.full_path(), sub=True, done=t.done, idt=4))
-        text.append(
-            '\n', _indent_text(4),
-            rclick(tr("mcd_task.done_task_button"), tr("mcd_task.done_task_hover"), "{} list-done",
-                   color=RColor.dark_gray)
-        )
+        for t in root.task_manager.split_sub_tasks_by_done()[0 if not done else 1]:  # type: Task
+            text.append('\n', _task_info_simple(t.full_path(), sub=True, done=t.done, desc=False, idt=4,
+                                                desc_text=t.description if t.description != '' else None))
+        if not done:
+            text.append(
+                '\n', _indent_text(4),
+                rclick(tr("mcd_task.done_task_button"), tr("mcd_task.done_task_hover"), "{} list-done".format(PREFIX),
+                       color=RColor.dark_gray)
+            )
     else:
         text.append(_task_info_flexible(target_task.full_path(), sub=False))
     return text
@@ -291,26 +320,30 @@ def illegal_call(source: CommandSource):
 
 # Infos
 def info_player(source: CommandSource, name: str) -> None:
-    tasks = root.task_manager.get_responsible_manager().get_player_tasks(name)
+    tasks = root.task_manager.get_responsible_manager()[name]
     text = RTextList(
-        RText(tr("mcd_task.player_tasks_title", name, str(len(tasks))), color=RColor.green, styles=RColor.bold),
+        RText(tr("mcd_task.player_tasks_title", name, str(len(tasks))), color=RColor.green, styles=RStyle.bold),
     )
     for t in tasks:
         tdone = root.task_manager[t].done
-        text.append('\n', _task_info_simple(t, done=tdone, idt=4, sub=False, desc=False))
+        text.append('\n', _task_info_simple(t, full=True, done=tdone, idt=4, sub=False))
 
     source.reply(text)
 
 
 def info_task(source: CommandSource, title: Optional[str] = None, prefix=None, done=False) -> None:
     # Task list
-    task_list = _info_task(source, title, done=done)
-    if task_list is None:
+    try:
+        task_list = _info_task(title, done=done)
+    except TaskNotFoundError:
+        task_not_found(source)
+        if DEBUG_MODE:
+            raise
         return
 
     # Prefix text
     if prefix is None:
-        prefix = tr("mcd_task.task_info_list_title" if title is None else "mcd_task.info_task_single_title")
+        prefix = tr("mcd_task.info_task_title" if title is None else "mcd_task.info_task_single_title")
     pref_text = RText(prefix, color=RColor.green, styles=RStyle.bold)
 
     add_button = _add_task_button(title)
@@ -322,55 +355,54 @@ def info_task(source: CommandSource, title: Optional[str] = None, prefix=None, d
 
 # Others
 def set_task_deadline(source: CommandSource, titles: str, ddl: str) -> None:
-    try:
-        target_task = root.task_manager[TitleList(titles)]
-    except TaskNotFoundError:
+    if not root.task_manager.exists(TitleList(titles)):
         task_not_found(source)
         return
-    try:
-        period = float(ddl) * 3600 * 24
-    except ValueError:
-        errcmd = "{} deadline {} ".format(PREFIX, titles)
-        source.reply(
-            rclick(tr("mcd_task.invalid_number"),
-                   tr("mcd_task.resuggest_cmd_hover", errcmd),
-                   errcmd, color=RColor.red)
+    deadline = float(time.time()) + float(ddl) * 3600 * 24
+    root.task_manager.set_deadline(TitleList(titles), deadline)
+    source.reply(
+        RText(
+            tr("mcd_task.ddl_set", titles,
+               ddl, formatted_time(deadline)),
+            color=RColor.yellow
         )
-    else:
-        deadline = float(time.time()) + period
-        target_task.set_deadline(deadline)
-        source.reply(
-            RText(tr("mcd_task.ddl_set",
-                     str(target_task.full_path()), ddl, formatted_time(deadline)), color=RColor.yellow)
-        )
+    )
+
+
+def list_task(source: CommandSource):
+    info_task(source)
 
 
 def show_help(source: CommandSource):
     meta = source.get_server().get_plugin_metadata('mcd_task')
     help_message = source.get_server().tr('mcd_task.help_msg').format(PREFIX, meta.name, str(meta.version))
-    help_msg_rtext = RTextList()
+    help_msg_rtext = ''
+    symbol = 0
     for line in help_message.splitlines():
+        if help_msg_rtext != '':
+            help_msg_rtext += '\n'
         result = re.search(r'(?<=§7){}[\S ]*?(?=§)'.format(PREFIX), line)
-        if result is not None:
+        if result is not None and symbol != 2:
             cmd = result.group() + ' '
-            help_msg_rtext.append(RText(line).c(RAction.suggest_command, cmd).h('点击以填入 §7{}§r'.format(cmd)))
+            help_msg_rtext += RText(line).c(RAction.suggest_command, cmd).h(
+                tr("mcd_task.help_msg_suggest_hover", cmd))
+            symbol = 1
         else:
-            help_msg_rtext.append(line)
-        if line != help_message.splitlines()[-1]:
-            help_msg_rtext.append('\n')
+            help_msg_rtext += line
+            if symbol == 1:
+                symbol += 1
     source.reply(help_msg_rtext)
 
 
 def add_task(source: CommandSource, titles: str, desc: str = ''):
     titles = TitleList(titles)
-    title = titles.pop_tail()
-    father = root.task_manager[title]
-    if title not in father.child_titles:
-        father.add_task(titles.removed, desc=desc)
+    titles_for_text = titles.copy()
+    if not root.task_manager.exists(titles):
+        root.task_manager.add_task(titles, desc=desc)
+        root.debug("Added task: {}".format(str(titles_for_text)))
+        info_task(source, title=titles_for_text.head, prefix=tr("mcd_task.new_task_created"))
     else:
         task_already_exist(source)
-        return
-    info_task(source, title, tr("mcd_task.new_task_created"))
 
 
 def all_tasks_detail(source: CommandSource):
@@ -383,34 +415,49 @@ def all_tasks_detail(source: CommandSource):
 
 
 def remove_task(source: CommandSource, titles: str):
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     root.task_manager.delete_task(TitleList(titles))
     source.reply(tr("mcd_task.deleted_task", "§e{}§r".format(titles)))
 
 
 def rename_task(source: CommandSource, old_titles: str, new_title: str) -> None:
+    if not root.task_manager.exists(TitleList(old_titles)):
+        task_not_found(source)
+        return
     if '.' in list(new_title):
         source.reply(tr("mcd_task.illegal_title_with_dot", new_title))
-    root.task_manager.rename_task(TitleList(old_titles))
+    root.task_manager.rename_task(TitleList(old_titles), new_title)
     new_titles = TitleList(old_titles)
     new_titles.pop_tail()
     new_titles.append(new_title)
-    info_task(source, str(new_titles))
+    info_task(source, prefix=str(new_titles))
 
 
 def edit_desc(source: CommandSource, titles: str, desc: str) -> None:
     title_list = TitleList(titles)
+    if not root.task_manager.exists(title_list):
+        task_not_found(source)
+        return
     root.task_manager.edit_desc(title_list, desc)
-    info_task(source, titles)
+    info_task(source, title=titles)
 
 
 def set_done(source: CommandSource, titles: str) -> None:
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     root.task_manager.done_task(TitleList(titles))
-    info_task(source, titles)
+    info_task(source, title=titles)
 
 
 def set_undone(source: CommandSource, titles: str) -> None:
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     root.task_manager.undone_task(TitleList(titles))
-    info_task(source, titles)
+    info_task(source, title=titles)
 
 
 def list_done(source: CommandSource):
@@ -418,34 +465,47 @@ def list_done(source: CommandSource):
 
 
 def set_responsible(source: CommandSource, titles: str, players: Optional[str] = None) -> None:
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     if players is None:
         if isinstance(source, PlayerCommandSource):
             players = source.player
         else:
             illegal_call(source)
-    num = root.task_manager.set_responsible(TitleList(titles), *players.split('.'))
+            return
+    players = players.split('.')
+    num = root.task_manager.set_responsible(TitleList(titles), *players)
     list_responsible(source, titles, prefix=tr("mcd_task.added_responsibles_title", num))
 
 
 def rm_responsible(source: CommandSource, titles: str, players: Optional[str] = None) -> None:
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     if players is None:
         if isinstance(source, PlayerCommandSource):
             players = source.player
         else:
             illegal_call(source)
-    num = root.task_manager.rm_responsible(TitleList(titles), *players.split('.'))
-    list_responsible(source, titles, player_removed=players,
+    players = players.split('.')
+    removed = root.task_manager.rm_responsible(TitleList(titles), *players)
+    num = len(removed)
+    list_responsible(source, titles, player_removed=removed,
                      prefix=tr("mcd_task.removed_responsibles_title", num))
 
 
 def list_responsible(source: CommandSource, titles: str, player_removed=None, prefix=None) -> None:
+    if not root.task_manager.exists(TitleList(titles)):
+        task_not_found(source)
+        return
     if player_removed is None:
         player_removed = []
     player_list = root.task_manager.get_responsible_manager().get_responsibles(titles)
     task_done = root.task_manager[titles].done
     text = RTextList(
-        tr("mcd_task.list_responsible_title") if prefix is None else prefix, '\n',
-        _task_info_simple(titles, full=True, sub=True, done=task_done, idt=4, desc=False)
+        RText(tr("mcd_task.list_responsible_title") if prefix is None else prefix, RColor.green, RStyle.bold),
+        '\n', _task_info_simple(titles, full=True, sub=True, done=task_done, idt=4)
     )
     for p in player_list:
         text.append(
@@ -468,7 +528,7 @@ def inherit_responsible(info: Info, old_name: str, new_name: str):
     info.get_server().tell(new_name, tr("mcd_task.on_player_renamed", num))
 
 
-def task_timed_out(server: PluginServerInterface, player: str, player_tasks: List[Task]):
+def task_timed_out(server: PluginServerInterface, player: str, player_tasks: List[TaskBase]):
     text = RTextList(RText(tr("mcd_task.on_player_joined", len(player_tasks)), color=RColor.red, styles=RStyle.bold))
     for t in player_tasks:
         text.append('\n', _task_info_simple(t.full_path(), full=True, sub=True, done=t.done, idt=4))
